+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ## Unreleased
 
-- Exclude assemblies that set DefinedTypes to null instead of an empty array
-  to prevent failures on GetTypes(). This issue occurred with the
-  Microsoft.WindowsAzure.Storage.dll assembly.
+- CertificateDsc.Common.psm1
+  - Exclude assemblies that set DefinedTypes to null instead of an empty array
+    to prevent failures on GetTypes(). This issue occurred with the
+    Microsoft.WindowsAzure.Storage.dll assembly.
 
 ## 4.4.0.0